requirements.txt
# Flask
# Flask-MySQLdb
# Flask-WTF
# WTForms
# mysqlclient
# bootstrap-flask
# Jinja2

# Flask==2.0.1
# Flask-MySQLdb==1.0.1
# Flask-WTF==0.15.1
# WTForms==2.3.3
# mysqlclient==2.1.1
# bootstrap-flask==2.0.1
# Jinja2==3.0.1

Flask==2.0.1
Flask-MySQLdb==1.0.1
Flask-WTF==0.15.1
WTForms==2.3.3
bootstrap-flask==2.0.1
Jinja2==3.0.1
email_validator==1.1.3
Flask-Login==0.5.0
werkzeug==2.0.1
python-dotenv==0.19.0
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
mysqlclient==2.0.3